Added a payload to the server pulse response that holds all relevant
game information, such as is the game active, who is the active player,
where the suspects and weapons are.  Added the ability to move between
rooms.  We need to work on the CLI to make it more user friendly as to
what room the user is currently in, and what are the rooms around them.
I believe Jason is working on creating some sort of ASCII map that will
make this a lot better.