Cleaned up some mis-labeled code.  Added foundations for suggest
functionality.  The server needs logic to pass the suggestion to each
client.  This could be done by adding a "toUuid" to the Messages and let
the client figure out if they should handle it, or by sending a message
directly to each client (doesn't exactly fit with the response model
currently in place since the client that is being sent the suggestion is
not the one who initiated it).