Added ability to make notes in the Notebook object.  It isn't very smart
at the moment and could be improved to maybe have two sections: marking
cards simply yes or no, and making free-form notes.  Added more logic
for when players disconnect, hopefully fixing some of the exceptions I
was seeing.  If a player disconnects before the game starts, that
suspect is free to be chosen again.  If a player disconnects after the
game starts, the server just ends the game and closes (for now).